backend/requirements.txt
fastapi
python-multipart
uvicorn
opencv-python
numpy
keras_facenet
scipy
transformers
torch
Pillow